webdjango/requirements/base.txt
# Base will provide the common requirements among all environments (development, live, staging, etc).

# Django
Django==2.1.4

#PostGree Daabase
psycopg2==2.7.6.1

# Enviroment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.9.0

# Django REST Framework JWT token
djangorestframework-jwt==1.11.0

# Django REST Framework JSON.API Format
djangorestframework-jsonapi==2.6.0

# Timezone
pytz==2018.4

# Windows's Interface to access  CA and CRL certificates.
wincertstore==0.2

# Access-Control-Allow-Origin headers
django-cors-headers==2.4.0

django-extensions==2.1.2

django-dirtyfields==1.3.1

django-filter==2.0.0
#django-url-filter==0.3.8 Not Working For now

# For the API Documentation Made Eazy
coreapi==2.3.3
drf-yasg==1.10.2
flex==6.13.2
swagger-spec-validator==2.4.0

# For the CDNS



# Country choices for use with forms, flag icons static files, and a country field for models.
django-countries==5.3.2


# prices fields
django-prices==1.0.4

# Easily use, manipulate, and store unit-aware measurement objects
django-measurement==3.1.1
